+++ dev/index.js
@@ -40,8 +40,6 @@ function BGraph(props) {
             bgrapher.initBgraph(testOnlyDots(1000, 10000));
         } else if (props.bgraphType == 'testEdges') {
             bgrapher.initBgraph(testDotsEdges(1000, 1000));
-        } else if (props.bgraphType == 'testLarge') {
-            bgrapher._initTestBgraphLarge(5000, 10000);
         }
 
         return bgrapher
@@ -155,9 +153,6 @@ function BGraphForm(props) {
             </button>
             <button className="bgraphFormSubmit" onClick={handleSubmit("testEdges")}>
                 Test edges
-            </button>
-            <button className="bgraphFormSubmit" onClick={handleSubmit("testLarge")}>
-                Large test image
             </button>
         </form>
     );